backend/requirement.txt
django==4.2
pymongo
django-cors-headers
pandas
openpyxl
google.generativeai 
certifi>=2017.4.17
pycryptodome
whitenoise
channels
daphne
opencv-python
Pillow
boto3
mediapipe
djangorestframework
channels-redis 
google-cloud-speech 
websockets
deepface
tf-keras
vertexai
google.genai 

